Fix propTypes from react-native leaking to web context
