+++ src/views/TabView/TabBarTop.js
@@ -12,7 +12,6 @@ import type {
   NavigationAction,
   NavigationScreenProp,
   NavigationState,
-  NavigationRoute,
   Style,
 } from '../../TypeDefinition';
